Moved ladder graph solver into core directory; moved BGL solvers into separate directory
